# server/openapi.yaml
openapi: 3.2.0
info:
  title: Receipt Budgeter API
  version: 1.0.0
  description: >-
    Minimal OpenAPI spec for the Receipt Budgeter service. Provides endpoints
    for retrieving budget information, analyzing receipt images (file upload),
    and submitting receipts (JSON transactions).
servers:
  - url: http://localhost:3000
    description: Local development server

paths:
  /status:
    get:
      summary: Get service status
      operationId: getStatus
      responses:
        '200':
          description: Service is running
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                required:
                  - status
              examples:
                example-1:
                  value:
                    status: "ok"
                    
  /getBudgetInformation:
    get:
      summary: Get budget categories and accounts
      operationId: getBudgetInformation
      responses:
        '200':
          description: Budget categories and accounts returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetInformationResponse'
              examples:
                example-1:
                  summary: Example response
                  value:
                    availableCategories:
                      - id: food
                        name: Food & Dining
                      - id: transport
                        name: Transportation
                    accounts:
                      - id: checking
                        name: Checking Account
                      - id: credit-card
                        name: Credit Card

  /analyzeReceipt:
    post:
      summary: Analyze an uploaded receipt image
      operationId: analyzeReceipt
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: |-
                    Receipt image file to analyze.
                    Allowed MIME types: image/png, image/jpeg, image/webp, image/heic, image/heif
              required:
                - file
            encoding:
              file:
                # Some tools accept a comma-separated list here; this documents allowed types
                contentType: "image/png, image/jpeg, image/webp, image/heic, image/heif"
      responses:
        '200':
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyzeReceiptResponse'


  /submitReceipt:
    post:
      summary: Submit parsed receipt transactions
      operationId: submitReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitReceiptRequest'
            examples:
                example-1:
                  summary: Submit a single receipt split across categories
                  value:
                    accountId: "checking"
                    merchantName: "Corner Grocery"
                    receiptDate: "2025-10-01"
                    subtotal: 42.5
                    tax: 3.4
                    total: 45.9
                    splits:
                      - categoryId: food
                        amount: 30.0
                        description: "Groceries"
                      - categoryId: household
                        amount: 15.9
                        description: "Household supplies (includes tax)"
      responses:
        '200':
          description: Receipt submitted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  id:
                    type: string
                required:
                  - ok
              examples:
                example-1:
                  value:
                    ok: true
                    id: "abc123"

components:
  schemas:
    Category:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name

    Account:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
      required:
        - id
        - name

    BudgetInformationResponse:
      type: object
      properties:
        availableCategories:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
      required:
        - availableCategories
        - accounts

    AnalyzeReceiptResponse:
      type: object
      properties:
        receiptData:
          $ref: '#/components/schemas/ReceiptData'
        items:
          type: array
          items:
            $ref: '#/components/schemas/AnalyzedItem'
      required:
        - receiptData
        - items

    ReceiptData:
      type: object
      properties:
        subtotal:
          type: number
        total:
          type: number
        storeName:
          type: string
        storeLocation:
          type: string
        taxAmount:
          type: number
        taxRates:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              rate: 
                type: number
              description:
                type: string
      required:
        - total

    AnalyzedItem:
      type: object
      properties:
        itemId:
          type: string
        itemName:
          type: string
        itemKind:
          type: string
        price:
          type: number
        budgetCategoryName:
          type: string
        budgetCategoryId:
          type: string
        taxesApplied:
          type: array
          items:
            type: array
            items:
              type: string
      required:
        - itemName
        - price

    SubItem:
      type: object
      properties:
        name:
          type: string
        itemKind:
          type: string
          description: Sub-item kind (e.g., "bottle deposit", "discount", etc.)
        price:
          type: number
          format: float
          description: Sub-item price
        isPositiveValue:
          type: boolean
          description: Whether this sub-item adds to (true) or subtracts from (false) the receipt total
      required:
        - name
        - price

    AnalyzeItem:
      type: object
      properties:
        itemName:
          type: string
        name:
          type: string
        itemKind:
          type: string
        price:
          type: number
        budgetCategoryGuess:
          type: string
      required:
        - price

    Transaction:
      type: object
      properties:
        name:
          type: string
        itemKind:
          type: string
        budgetCategoryGuess:
          type: string
        price:
          type: number
          format: float
      required:
        - name
        - price
    Split:
      type: object
      properties:
        categoryId:
          type: string
          description: Budget category id (should match an available category)
        amount:
          type: number
          format: float
          description: Amount (including tax portion) allocated to this category
        description:
          type: string
          description: Optional free-text description of the split
      required:
        - categoryId
        - amount

    SubmitReceiptRequest:
      type: object
      properties:
        accountId:
          type: string
          description: Id of the account this multi-category transaction belongs to
        merchantName:
          type: string
        receiptDate:
          type: string
          format: date
        subtotal:
          type: number
          format: float
        tax:
          type: number
          format: float
        total:
          type: number
          format: float
        splits:
          type: array
          description: List of how the total cost is split across categories. Must sum to subtotal.
          items:
            $ref: '#/components/schemas/Split'
        doBankSync:
          type: boolean
          description: Whether to sync Actual Budget transactions with the bank before submitting this receipt to Actual Budget
          default: true
        createTransactions:
          type: boolean
          description: Whether to create a new transaction for this receipt
          default: false
      required:
        - accountId
        - splits

  securitySchemes: {}